Remove unused devise files, stripout unused features from devise setup
