+++ app/controllers/home_controller.rb
@@ -12,23 +12,4 @@ class HomeController < ApplicationController
 
     @user = current_user
   end
-
-  def profile
-    @user = current_user
-  end
-
-  def update_profile
-    @user = current_user
-    if @user.update(user_params)
-      redirect_to profile_path, notice: "Profile updated successfully."
-    else
-      render :profile
-    end
-  end
-
-private
-
-  def user_params
-    params.require(:user).permit(:image)
-  end
 end
