--- app/views/devise/mailer/password_change.html.erb
@@ -1,3 +1,0 @@
-<p>Hello <%= @resource.email %>!</p>
-
-<p>We're contacting you to notify you that your password has been changed.</p>
