+++ app/views/devise/registrations/edit.html.erb
@@ -1,3 +1,5 @@
+<%= link_to "Back", :back %>
+<br><br>
 <h2>Edit <%= resource_name.to_s.humanize %></h2>
 
 <%= form_for(resource, as: resource_name, url: registration_path(resource_name), html: { method: :put }) do |f| %>
@@ -31,24 +33,19 @@
     <%= f.password_field :current_password, autocomplete: "current-password" %>
   </div>
 
-
   <div class="field">
     <%= f.label :first_name %> <br />
     <%= f.text_field :first_name %>
   </div>
 
-    <div class="field">
+  <div class="field">
     <%= f.label :last_name %> <br />
     <%= f.text_field :last_name %>
   </div>
+  <br>
 
   <div class="actions">
     <%= f.submit "Update" %>
   </div>
 <% end %>
-
-<h3>Cancel my account</h3>
-
-<div>Unhappy? <%= button_to "Cancel my account", registration_path(resource_name), data: { confirm: "Are you sure?", turbo_confirm: "Are you sure?" }, method: :delete %></div>
-
-<%= link_to "Back", :back %>
+<br>
