+++ app/views/devise/sessions/new.html.erb
@@ -11,18 +11,10 @@
       <%= f.label :password %><br />
       <%= f.password_field :password, autocomplete: "current-password", class: "form-control"  %>
     </div>
-
-    <% if devise_mapping.rememberable? %>
-      <div class="field">
-        <%= f.check_box :remember_me %>
-        <%= f.label :remember_me %>
-      </div>
-    <% end %>
+    <br>
 
     <div class="actions">
       <%= f.submit "Log in" %>
     </div>
   </div>
 <% end %>
-
-<%= render "devise/shared/links" %>
