+++ config/routes.rb
@@ -1,9 +1,13 @@
 Rails.application.routes.draw do
-  devise_for :users
+  devise_for :users, controllers: {
+    registrations: 'users/registrations'
+  }
+
   root 'home#homepage'
   get 'homepage', to: 'home#homepage'
-  get 'profile', to: 'home#profile'
-  patch 'profile', to: 'home#update_profile'
+
+  get 'profile', to: 'profiles#profile', as: 'profile'
+  patch 'profile', to: 'profiles#update_avatar'
 
   resources :books, only: [:show, :index] do
     resources :reviews, only: [:new, :create]
@@ -11,15 +15,3 @@ Rails.application.routes.draw do
 
   resources :user_books, only: [:create]
 end
-
-
-# Some setup you must do manually if you haven't yet:
-
-#   Ensure you have overridden routes for generated controllers in your routes.rb.
-#   For example:
-
-#     Rails.application.routes.draw do
-#       devise_for :users, controllers: {
-#         sessions: 'users/sessions'
-#       }
-#     end
